interoperability/foundry/agents/stay/requirements.txt
# Stay Agent Dependencies - Microsoft Agent Framework Hosted Agent
#
# Design doc references:
# - Appendix A.2 lines 1544-1551: Installation packages for hosted agents
# - Line 1548: azure-ai-agentserver-agentframework for Agent Framework agents

# Azure AI hosted agent packages (preview)
azure-ai-projects>=2.0.0b2
azure-ai-agentserver-core
azure-ai-agentserver-agentframework

# Azure authentication
azure-identity

# Microsoft Agent Framework (used by existing agent logic)
agent-framework
agent-framework-azure-ai

# Required by existing agent logic
pydantic>=2.0.0
python-dotenv

# OpenAI for model access (used by Agent Framework)
openai>=1.0.0
